Changed weapon damage to a float. Changed beam weapon to deal continuous rather than discrete damage. Fixed bug with projectiles and open doors.

+++ EscapeFromEnceladus/EscapeFromEnceladus/Weapon/IWeapon.cs
@@ -11,6 +11,6 @@ namespace Enceladus.Weapon {
         /// <summary>
         /// Returns the base damage for this projectile.
         /// </summary>
-        int BaseDamage { get; }
+        float BaseDamage { get; }
     }
 }
